+++ email/email-rest/src/main/java/com/junbo/email/rest/resource/EmailTemplateResourceImpl.java
@@ -5,6 +5,8 @@
  */
 package com.junbo.email.rest.resource;
 
+import com.junbo.authorization.AuthorizeContext;
+import com.junbo.authorization.spec.error.AppErrors;
 import com.junbo.common.id.EmailTemplateId;
 import com.junbo.common.model.Results;
 import com.junbo.email.core.EmailTemplateService;
@@ -26,25 +28,46 @@ import javax.ws.rs.ext.Provider;
 @Component
 @Scope("prototype")
 public class EmailTemplateResourceImpl implements EmailTemplateResource{
+    private static final String EMAIL_ADMIN_SCOPE = "email.admin";
+    private static final String EMAIL_SERVICE_SCOPE = "email.service";
     @Autowired
     private EmailTemplateService templateService;
 
     public Promise<EmailTemplate> postEmailTemplate(EmailTemplate template) {
+        if (!AuthorizeContext.hasScopes(EMAIL_ADMIN_SCOPE)) {
+            throw AppErrors.INSTANCE.insufficientScope().exception();
+        }
+
         return templateService.postEmailTemplate(template);
     }
     public Promise<EmailTemplate> getEmailTemplate(EmailTemplateId id) {
+        if (!AuthorizeContext.hasScopes(EMAIL_ADMIN_SCOPE) && !AuthorizeContext.hasScopes(EMAIL_SERVICE_SCOPE)) {
+            throw AppErrors.INSTANCE.insufficientScope().exception();
+        }
+
         return templateService.getEmailTemplate(id.getValue());
     }
     public Promise<EmailTemplate> putEmailTemplate(EmailTemplateId id, EmailTemplate template) {
+        if (!AuthorizeContext.hasScopes(EMAIL_ADMIN_SCOPE)) {
+            throw AppErrors.INSTANCE.insufficientScope().exception();
+        }
+
         return templateService.putEmailTemplate(id.getValue(), template);
     }
     public Promise<Response> deleteEmailTemplate(EmailTemplateId id) {
+        if (!AuthorizeContext.hasScopes(EMAIL_ADMIN_SCOPE)) {
+            throw AppErrors.INSTANCE.insufficientScope().exception();
+        }
+
         templateService.deleteEmailTemplate(id.getValue());
         return Promise.pure(Response.status(204).build());
     }
 
     public Promise<Results<EmailTemplate>> getEmailTemplates(QueryParam queryParam) {
+        if (!AuthorizeContext.hasScopes(EMAIL_ADMIN_SCOPE) && !AuthorizeContext.hasScopes(EMAIL_SERVICE_SCOPE)) {
+            throw AppErrors.INSTANCE.insufficientScope().exception();
+        }
+
         return templateService.getEmailTemplates(queryParam);
     }
-
 }
